Moved keywords to Accounts and Contacts settings page
