+++ robot/EDA/resources/ContactsPageObject.py
@@ -102,99 +102,6 @@ class ContactDetailPage(DetailPage):
         if capture_screen:
             self.selenium.capture_page_screenshot()
 
-    def set_the_disable_preferred_phone_enforcement(self):
-        """ Verify that Disable Preferred Phone enforcement checkbox is checked.
-            Set the checkbox if it is not already set.
-            Do nothing if the checkbox is already checked.
-        """
-        # Checkbox for 'Disable Preferred Phone enforcement' needs to be marked as checked
-        if self.eda._check_if_element_exists(contacts_locators["preferred_phone_active"]):
-            self.builtin.log("Disable Preferred Phone enforcement is checked.")
-            return
-        else: 
-            self.builtin.log(
-                "Disable Preferred Phone enforcement is NOT checked.\n" +
-                "Opening EDIT mode"
-            )
-            self.selenium.click_button("Edit")
-            self.builtin.log("Setting 'Disable Preferred Phone enforcement' checkbox.")
-
-            self.selenium.wait_until_page_contains_element(
-                contacts_locators["disable_checked"],
-                timeout=60,
-            )
-            self.selenium.driver.execute_script(
-                "arguments[0].click()", 
-                self.selenium.driver.find_element_by_xpath(contacts_locators["disable_checked"])
-            )
-            self.selenium.click_button("Save")
-            self.eda.close_toast_message()
-            self.builtin.log(
-                "Disable Preferred Phone enforcement checkbox has been checked.\n" +
-                "Saving changes.\n" +
-                "Proper configuration is in place for testing 'Disable Preferred Phone enforcement'."
-            )
-
-    def clear_the_disable_preferred_phone_enforcement(self):
-        """ Verify that Disable Preferred Phone enforcement checkbox is not checked 
-            Clear the checkbox if it is set
-            Do nothing if the checkbox is already empty
-        """
-
-        self.selenium.wait_until_page_contains_element(
-            contacts_locators["accounts_contacts"],
-            timeout=60
-        )
-        self.open_item(
-            contacts_locators["accounts_contacts"],
-            "Cannot find Account and Contacts on EDA Settings page", 
-            True
-        )
-
-        # Checkbox for 'Disable Preferred Phone enforcement' should be empty
-        if self.eda._check_if_element_exists(contacts_locators["disable_preferred_phone"]):
-            self.builtin.log("Disable Preferred Phone enforcement is clear.")
-        else: 
-
-            # Click on Edit button
-            self.selenium.click_button("Edit")
-            self.builtin.log("Setting 'Disable Preferred Phone enforcement' checkbox.")
-            self.builtin.log(
-                "Disable Preferred Phone enforcement is currently checked.\n" +
-                "In EDIT mode"
-            )
-            self.builtin.log("Clearing 'Disable Preferred Phone enforcement' checkbox.")
-            self.selenium.driver.execute_script(
-                "arguments[0].click()", 
-                self.selenium.driver.find_element_by_xpath(contacts_locators["disable_checked"])
-            )
-            self.selenium.click_button("Save")
-            self.eda.close_toast_message()
-
-        # Before running the 'Run Cleanup', we need to choose a 
-        # field for the Preferred Phone to copy the value from, 
-        # if you're not using the standard Phone field (or it has
-        # no value) and more than one 'Phone' field is defined.
-        # We'll choose 'Work Phone'
-
-        self.selenium.click_button("Edit")
-
-        if self.eda._check_if_element_exists(contacts_locators["copy_from"]):
-            self.selenium.driver.execute_script(
-                "arguments[0].click()",
-                self.selenium.driver.find_element_by_xpath(contacts_locators["copy_from"])
-            )
-        self.selenium.get_webelement(contacts_locators["copy_from"]).send_keys("H" + Keys.ENTER)
-
-        # Click Save
-        self.selenium.click_button("Save")
-        self.eda.close_toast_message()
-        self.builtin.log(
-            "Disable Preferred Phone enforcement checkbox has been cleared.\n" +
-            "Saving changes.\n" +
-            "Proper configuration is in place for testing 'Disable Preferred Phone enforcement'."
-        )
-
     def run_phone_cleanup(self):
         """ Click on the 'Run Phone Cleanup' button.
             Also verifies that the 'Run Phone Cleanup' was queued to run.
@@ -352,83 +259,6 @@ class ContactDetailPage(DetailPage):
 
         self.selenium.click_element(contacts_locators["button_save_affiliation"])
         self.eda.close_toast_message()
-
-    def enable_enchanced_checkbox(self):
-        """ Ensure that the Enhanced Preferred Phone Functionality checkbox is checked 
-            Set the checkbox if it is not set
-            Do nothing if the checkbox is already set
-        """
-        # Checkbox for 'Disable Preferred Phone enforcement' needs to be marked as checked
-        if self.eda._check_if_element_exists(contacts_locators["enhanced_preferred_set"]):
-            self.builtin.log("Enable Enhanced Preferred Phone Functionality is checked.")
-            return
-        else: 
-            self.builtin.log(
-                "Enable Enhanced Preferred Phone Functionality is NOT checked.\n" +
-                "Opening EDIT mode"
-            )
-            self.selenium.click_button("Edit")
-            self.builtin.log("Setting 'Enable Enhanced Preferred Phone Functionality' checkbox.")
-            self.selenium.wait_until_page_contains_element(
-                contacts_locators["enhanced_preferred_clear_faux"],
-                timeout=60
-            )
-            self.selenium.driver.execute_script(
-                "arguments[0].click()", 
-                self.selenium.driver.find_element_by_xpath(contacts_locators["enhanced_preferred_clear_faux"])
-            )
-            self.selenium.click_button("Save")
-            self.eda.close_toast_message()
-            self.builtin.log(
-                "Enable Enhanced Preferred Phone Functionality setting has been set.\n" +
-                "Saving changes.\n" +
-                "Proper configuration is in place for testing 'Disable Preferred Phone enforcement'."
-            )
-
-    def disable_enchanced_checkbox(self):
-        """ Verify that Enable Enhanced Preferred Phone Functionality checkbox is not checked 
-            Clear the checkbox if it is set
-            Do nothing if the checkbox is already empty
-        """
-
-        self.selenium.wait_until_page_contains_element(
-            contacts_locators["accounts_contacts"],
-            timeout=60
-        )
-        self.open_item(
-            contacts_locators["accounts_contacts"],
-            "Cannot find Account and Contacts on EDA Settings page", 
-            True
-        )
-
-        # Click on Edit button
-        self.selenium.click_button("Edit")
-        self.builtin.log("Clearing 'Enable Enhanced Preferred Phone Functionality' checkbox.")
-
-        # Checkbox for 'Enable Enhanced Preferred Phone Functionality' should be empty
-        if self.eda._check_if_element_exists(contacts_locators["enhanced_preferred_clear"]):
-            self.builtin.log("Enable Enhanced Preferred Phone Functionality is clear.")
-            return
-        else: 
-            self.builtin.log(
-                "Enable Enhanced Preferred Phone Functionality is checked.\n" +
-                "Opening EDIT mode"
-            )
-            self.builtin.log("Clearing 'Enable Enhanced Preferred Phone Functionality' checkbox.")
-
-            # Clear the check in the checkbox
-            self.selenium.driver.execute_script(
-                "arguments[0].click()", 
-                self.selenium.driver.find_element_by_xpath(contacts_locators["enhanced_preferred_set_faux"])
-            )
-        # Click Save
-        self.selenium.click_button("Save")
-        self.eda.close_toast_message()
-        self.builtin.log(
-            "Enable Enhanced Preferred Phone Functionality setting has been cleared.\n" +
-            "Saving changes.\n" +
-            "Proper configuration is in place for testing 'Disable Preferred Phone enforcement'."
-        )
 
     def wait_and_refresh_static_page_until_text(self, search_text, wait_time, loc_frame, loc_text):
         """ Wait for text to appear on static page.  Note that the page is refreshed each 'wait_time' until
