+++ robot/EDA/resources/locators.py
@@ -165,6 +165,16 @@ eda_lex_locators = {
         "checkbox": '//div[contains(@class,"uiInputCheckbox")]/label/span[text()="{}"]/../following-sibling::input[@type="checkbox"]',
         "save": "//div[contains(@class, 'footer') or contains(@class, 'Footer')]/descendant::button[@title='Save']",
     },
+    "accounts_contacts_settings_locators": {
+        "copy_from": "//select[@class='contact-preferred-phone-picklist-input-select select uiInput uiInputSelect uiInput--default uiInput--select']",
+        "disable_checked": "(//span[text()='Disable Preferred Phone enforcement:']/following::div/div/div/label/input/following-sibling::span)[1]",
+        "disable_preferred_phone": "//div/span[text()='Disable Preferred Phone enforcement:']/following::div[1]/div/div/label/span/img[@alt='False']",
+        "enhanced_preferred_clear": "//div/span[text()='Enable Enhanced Preferred Phone Functionality:']/following::div[1]/div/div/label/span/img[@alt='False']",
+        "enhanced_preferred_clear_faux": "//span[text()='Enable Enhanced Preferred Phone Functionality:']/following::div[1]/div/div/label/input/following::span[1]",
+        "enhanced_preferred_set": "//span[text()='Enable Enhanced Preferred Phone Functionality:']/following::div[1]/div/div/label/span/img[@alt='True']",
+        "enhanced_preferred_set_faux": "//span[text()='Enable Enhanced Preferred Phone Functionality:']/following::div[1]/div/div/label/input/following::span[1]",
+        "preferred_phone_active": "//div/span[text()='Disable Preferred Phone enforcement:']/following::div[1]/div/div/label/span/img[@alt='True']",
+    },
     "relationships_settings_locators": {
         "sub_tab": "//div[@id='relTabs']/descendant::li[contains(@class, 'slds-text-heading--label')]/a[text()='{}']",
     }
@@ -183,21 +193,13 @@ contacts_locators = {
     "field_for_work_phone": "//div//label//span[contains(text(),'Work Phone')]/../following-sibling::input",
     "which_footer_cancel": "//div[contains(@class,'footer')]/button[@title='Cancel']//span[text()='Cancel']",
     "footer_save": "//div[contains(@class,'modal-footer')]//span[text()='Save']",
-    "disable_preferred_phone": "//div/span[text()='Disable Preferred Phone enforcement:']/following::div[1]/div/div/label/span/img[@alt='False']",
-    "preferred_phone_active": "//div/span[text()='Disable Preferred Phone enforcement:']/following::div[1]/div/div/label/span/img[@alt='True']",
-    "disable_checked": "(//span[text()='Disable Preferred Phone enforcement:']/following::div/div/div/label/input/following-sibling::span)[1]",
     "accounts_contacts": "//a[contains(text(),'Accounts and Contacts')]",
     "details_tab": "//div[contains(@class,'normal')]//span[@class='title' and text()='Details']",
     "phone_home": "//span[text()='Home Phone']/../following-sibling::input",
     "run_cleanup": "//button[text()='Run Cleanup']",
     "phone_verify": "//div//span[text()='Home Phone']/../following-sibling::div//span//span[text()='123-123-1234']",
     "home_phone_verify": "//span[text()='Home Phone']/../following::div//span//span[text()='123-123-1234']",
-    "copy_from": "//select[@class='contact-preferred-phone-picklist-input-select select uiInput uiInputSelect uiInput--default uiInput--select']",
     "successful_run": "//span[text()='The process was queued successfully. An email will be sent at the completion of the job.']",
-    "enhanced_preferred_clear": "//div/span[text()='Enable Enhanced Preferred Phone Functionality:']/following::div[1]/div/div/label/span/img[@alt='False']",
-    "enhanced_preferred_clear_faux": "//span[text()='Enable Enhanced Preferred Phone Functionality:']/following::div[1]/div/div/label/input/following::span[1]",
-    "enhanced_preferred_set": "//span[text()='Enable Enhanced Preferred Phone Functionality:']/following::div[1]/div/div/label/span/img[@alt='True']",
-    "enhanced_preferred_set_faux": "//span[text()='Enable Enhanced Preferred Phone Functionality:']/following::div[1]/div/div/label/input/following::span[1]",
     "apex_jobs": "//a/mark[text()='{}']",
     "primary_business_organization": "(//span[text()='Primary Business Organization']/following::div/div/div/div/input[@title='Search Accounts'])[1]",
     "button_save_affiliation": "//button[@title='Save']//span[text()='Save']",
